Split code into smaller files
Fixes #18
