+++ src/game/game_result.rs
@@ -1,0 +1,6 @@
+use super::piece::Piece;
+
+pub enum GameResult {
+    Win(Piece),
+    Tie,
+}
