+++ src/game/mod.rs
@@ -1,0 +1,5 @@
+pub mod board;
+pub mod field;
+pub mod game;
+pub mod game_result;
+pub mod piece;
