+++ src/player/mod.rs
@@ -1,0 +1,3 @@
+pub mod ai_player;
+pub mod local_player;
+pub mod player;
